+++ packages/types/src/span.ts
@@ -1,5 +1,5 @@
-/** SpanContext holding trace_id, span_id */
-export interface SpanContext {
+/** Span holding trace_id, span_id */
+export interface Span {
   /** Return a traceparent compatible header string */
   toTraceparent(): string;
   /** Convert the object to JSON */