+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - [opentracing] ref: Removed opentracing package
 - [integrations] feat: Add tracing integration
-- [hub] feat: Add tracing related function to scope and hub (`Scope.startSpan`, `Scope.setSpanContext`, `Hub.traceHeaders`)
+- [hub] feat: Add tracing related function to scope and hub (`Scope.startSpan`, `Scope.setSpan`, `Hub.traceHeaders`)
 - [hub] feat: Add new function to Scope `setContext`
 - [hub] feat: Add new function to Scope `setTransaction`